+++ src/resources/index.ts
@@ -4,6 +4,7 @@ export * from './shared';
 export { APIKeys } from './api-keys';
 export { AuditLogs } from './audit-logs';
 export { Compute } from './compute/compute';
+export { NKS } from './nks/nks';
 export { Networking } from './networking/networking';
 export { Operations } from './operations';
 export { Organizations } from './organizations';