+++ src/resources/nks/clusters/pools/index.ts
@@ -1,0 +1,23 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+export { Availability, type AvailabilityCreateParams, type AvailabilityUpdateParams } from './availability';
+export {
+  Nodes,
+  type NKSNode,
+  type NKSNodeList,
+  type NodeListParams,
+  type NodeGetParams,
+  type NKSNodesCursor,
+} from './nodes/index';
+export {
+  Pools,
+  type NKSNodeConfig,
+  type NKSNodePool,
+  type NKSNodePoolList,
+  type PoolCreateParams,
+  type PoolUpdateParams,
+  type PoolListParams,
+  type PoolDeleteParams,
+  type PoolGetParams,
+  type NKSNodePoolsCursor,
+} from './pools';