+++ tests/api-resources/nks/clusters/pools/nodes/volumes.test.ts
@@ -1,0 +1,60 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+import NirvanaLabs from '@nirvana-labs/nirvana';
+
+const client = new NirvanaLabs({
+  apiKey: 'My API Key',
+  baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
+});
+
+describe('resource volumes', () => {
+  // Mock server tests are disabled
+  test.skip('list: only required params', async () => {
+    const responsePromise = client.nks.clusters.pools.nodes.volumes.list('node_id', {
+      cluster_id: 'cluster_id',
+      pool_id: 'pool_id',
+    });
+    const rawResponse = await responsePromise.asResponse();
+    expect(rawResponse).toBeInstanceOf(Response);
+    const response = await responsePromise;
+    expect(response).not.toBeInstanceOf(Response);
+    const dataAndResponse = await responsePromise.withResponse();
+    expect(dataAndResponse.data).toBe(response);
+    expect(dataAndResponse.response).toBe(rawResponse);
+  });
+
+  // Mock server tests are disabled
+  test.skip('list: required and optional params', async () => {
+    const response = await client.nks.clusters.pools.nodes.volumes.list('node_id', {
+      cluster_id: 'cluster_id',
+      pool_id: 'pool_id',
+      cursor: 'cursor',
+      limit: 10,
+    });
+  });
+
+  // Mock server tests are disabled
+  test.skip('get: only required params', async () => {
+    const responsePromise = client.nks.clusters.pools.nodes.volumes.get('volume_id', {
+      cluster_id: 'cluster_id',
+      pool_id: 'pool_id',
+      node_id: 'node_id',
+    });
+    const rawResponse = await responsePromise.asResponse();
+    expect(rawResponse).toBeInstanceOf(Response);
+    const response = await responsePromise;
+    expect(response).not.toBeInstanceOf(Response);
+    const dataAndResponse = await responsePromise.withResponse();
+    expect(dataAndResponse.data).toBe(response);
+    expect(dataAndResponse.response).toBe(rawResponse);
+  });
+
+  // Mock server tests are disabled
+  test.skip('get: required and optional params', async () => {
+    const response = await client.nks.clusters.pools.nodes.volumes.get('volume_id', {
+      cluster_id: 'cluster_id',
+      pool_id: 'pool_id',
+      node_id: 'node_id',
+    });
+  });
+});